refactor(engine): move buildCustomElementConstructor to engine-dom
