+++ packages/@lwc/engine/dom/src/apis/build-custom-element-constructor.ts
@@ -5,7 +5,18 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 
+import { create, isUndefined, keys } from '@lwc/shared';
+
 import { LightningElement } from './lightning-element';
+
+import {
+    createVM,
+    connectRootElement,
+    disconnectedRootElement,
+    getAttrNameFromPropName,
+    getComponentInternalDef,
+    isAttributeLocked,
+} from '../../../src';
 
 type ComponentConstructor = typeof LightningElement;
 type HTMLElementConstructor = typeof HTMLElement;
@@ -37,4 +48,57 @@ export function deprecatedBuildCustomElementConstructor(
     }
 
     return Ctor.CustomElementConstructor;
+}
+
+export function buildCustomElementConstructor(Ctor: ComponentConstructor): HTMLElementConstructor {
+    const def = getComponentInternalDef(Ctor);
+
+    // generating the hash table for attributes to avoid duplicate fields and facilitate validation
+    // and false positives in case of inheritance.
+    const attributeToPropMap: Record<string, string> = create(null);
+    for (const propName in def.props) {
+        attributeToPropMap[getAttrNameFromPropName(propName)] = propName;
+    }
+
+    return class extends def.bridge {
+        constructor() {
+            super();
+            createVM(this, def, {
+                mode: 'open',
+                isRoot: true,
+                owner: null,
+                tagName: this.tagName,
+            });
+        }
+        connectedCallback() {
+            connectRootElement(this);
+        }
+        disconnectedCallback() {
+            disconnectedRootElement(this);
+        }
+        attributeChangedCallback(attrName: string, oldValue: string, newValue: string) {
+            if (oldValue === newValue) {
+                // Ignore same values.
+                return;
+            }
+            const propName = attributeToPropMap[attrName];
+            if (isUndefined(propName)) {
+                // Ignore unknown attributes.
+                return;
+            }
+            if (!isAttributeLocked(this, attrName)) {
+                // Ignore changes triggered by the engine itself during:
+                // * diffing when public props are attempting to reflect to the DOM
+                // * component via `this.setAttribute()`, should never update the prop.
+                // Both cases, the the setAttribute call is always wrap by the unlocking of the
+                // attribute to be changed
+                return;
+            }
+            // Reflect attribute change to the corresponding property when changed from outside.
+            (this as any)[propName] = newValue;
+        }
+        // Specify which attributes for which we want to reflect changes back to their corresponding
+        // properties via attributeChangedCallback.
+        static observedAttributes = keys(attributeToPropMap);
+    };
 }
