+++ packages/@lwc/engine/dom/src/apis/lightning-element.ts
@@ -6,7 +6,9 @@
  */
 import { defineProperty, freeze, isUndefined, seal } from '@lwc/shared';
 
-import { buildCustomElementConstructor, LightningElement } from '../../../src';
+import { buildCustomElementConstructor } from './build-custom-element-constructor';
+
+import { LightningElement } from '../../../src';
 
 type ComponentConstructor = typeof LightningElement;
 type HTMLElementConstructor = typeof HTMLElement;
