+++ packages/@lwc/engine/src/framework/main.ts
@@ -19,6 +19,15 @@ export { default as api } from './decorators/api';
 export { default as track } from './decorators/track';
 export { default as wire } from './decorators/wire';
 export { readonly } from './readonly';
-export { buildCustomElementConstructor } from './wc';
 
 export { setFeatureFlag, setFeatureFlagForTest } from '@lwc/features';
+
+// Internal APIs -----------------------------------------------------------------------------------
+export { getAttrNameFromPropName, isAttributeLocked } from './attributes';
+export { getComponentInternalDef, setElementProto } from './def';
+export {
+    createVM,
+    connectRootElement,
+    disconnectedRootElement,
+    getAssociatedVMIfPresent,
+} from './vm';
